Fix the same int32 overflow pattern in other Triton kernels
Follow-up audit after fixing the int32 pointer-arithmetic overflow in
_chunk_cumsum_fwd_kernel/_chunk_cumsum_bwd_kernel (previous commit):
the same unguarded `pid_* * chunk_size * stride_*_seqlen`-style chained
multiplication, with an uncast 32-bit leading operand, exists in every
other pid_bc/pid_c/pid_ch-indexed kernel across ssd_chunk_state.py (the
remaining kernels: _chunk_state_fwd_kernel, _chunk_state_bwd_dx_kernel,
_chunk_state_bwd_db_kernel, _chunk_state_bwd_ddAcs_stable_kernel, and
_chunk_state_varlen_kernel), all eleven kernels in ssd_chunk_scan.py,
the fused kernel in ssd_combined.py, and both kernels in ssd_bmm.py.
Same fix in each case: cast the leading program-id-derived index to
tl.int64 immediately after it's obtained.

Also found a related but distinct instance of the same overflow class
in layer_norm.py, layernorm_gated.py, and k_activations.py: these
kernels compute `row * stride_x_row` where `row = tl.program_id(0)`
ranges over all flattened rows (batch*seqlen), and stride_x_row is the
row's width (hidden dim). Unlike the chunk-indexed kernels, this
doesn't need a non-contiguous view to overflow -- an ordinarily
contiguous (M, N) input is enough once M * N exceeds 2**31 - 1, since
`row` alone ranges up to M-1. Same fix: cast `row`/`row_block_id` to
tl.int64 right after tl.program_id().

All of these were found by grepping for the same unguarded pattern
after the first fix, then confirming each finding is a *real* crash
(not just a theoretical risk) before fixing it:

- mamba_chunk_scan_combined, called with `x` sliced non-contiguously
  out of a wide parent tensor (same shape as the original bug report),
  crashed with the identical "Triton Error [CUDA]: an illegal memory
  access was encountered" pre-fix, tracing into
  _chunk_state_fwd_kernel -- a kernel not touched by the previous fix.
  Passes post-fix. Added as
  test_mamba_chunk_scan_combined_noncontiguous_wide_view_no_overflow
  in tests/ops/triton/test_ssd.py; exercises the ssd_chunk_state.py,
  ssd_chunk_scan.py, ssd_combined.py, and ssd_bmm.py fixes together.
- layernorm_fn (RMSNorm path) on a contiguous (115_866, 18_560) input
  crashed identically pre-fix, passes post-fix. Added as
  test_layer_norm_gated_large_row_count_no_overflow in
  tests/ops/triton/test_layernorm_gated.py.

Not fixed here, flagged for follow-up:
- ssd_state_passing.py has `(nchunks - 1) * stride_*_chunk` with an
  uncast nchunks, structurally the same class of risk, but nchunks
  isn't program-id-derived and stride_*_chunk is a state-tensor stride
  (nheads * headdim-ish) rather than a wide/sliceable seqlen stride --
  no concrete overflow scenario constructed, so left as a flag rather
  than a speculative fix.
- selective_state_update.py only multiplies pid_b/pid_h against
  batch/head strides for single-token inference state updates --
  lower risk, not touched.
- mamba_ssm/ops/triton/mamba3/ (Mamba-3, a separate newer subsystem)
  uses pid_chunk * stride_*_chunk as a single multiplication against a
  chunk-dim stride, not a chained product through a wide seqlen
  stride -- different shape of risk, not audited in depth here.

+++ mamba_ssm/ops/triton/k_activations.py
@@ -31,7 +31,9 @@ def _swiglu_fwd_kernel(
     BLOCK_N: tl.constexpr,
 ):
     # Map the program id to the row of X and Y it should compute.
-    row = tl.program_id(0)
+    # if row * stride_x_row is large, may overflow int32, so use 64 bit
+    # https://github.com/triton-lang/triton/issues/1058
+    row = tl.program_id(0).to(tl.int64)
     start_col = tl.program_id(1) * BLOCK_N
     X += row * stride_x_row
     Y += row * stride_y_row
@@ -93,7 +95,9 @@ def _swiglu_bwd_kernel(
     RECOMPUTE_OUTPUT: tl.constexpr,
 ):
     # Map the program id to the row of X and Y it should compute.
-    row = tl.program_id(0)
+    # if row * stride_x_row is large, may overflow int32, so use 64 bit
+    # https://github.com/triton-lang/triton/issues/1058
+    row = tl.program_id(0).to(tl.int64)
     start_col = tl.program_id(1) * BLOCK_N
     X += row * stride_x_row
     Y += row * stride_y_row
