+++ mamba_ssm/ops/triton/layernorm_gated.py
@@ -63,7 +63,9 @@ def _layer_norm_fwd_1pass_kernel(
     IS_RMS_NORM: tl.constexpr,
 ):
     # Map the program id to the row of X and Y it should compute.
-    row = tl.program_id(0)
+    # if row * stride_x_row is large, may overflow int32, so use 64 bit
+    # https://github.com/triton-lang/triton/issues/1058
+    row = tl.program_id(0).to(tl.int64)
     group = tl.program_id(1)
     X += row * stride_x_row + group * N
     Y += row * stride_y_row + group * N
@@ -185,7 +187,9 @@ def _layer_norm_bwd_kernel(
     BLOCK_N: tl.constexpr,
 ):
     # Map the program id to the elements of X, DX, and DY it should compute.
-    row_block_id = tl.program_id(0)
+    # if row_start * stride_x_row is large, may overflow int32, so use 64 bit
+    # https://github.com/triton-lang/triton/issues/1058
+    row_block_id = tl.program_id(0).to(tl.int64)
     group = tl.program_id(1)
     row_start = row_block_id * rows_per_program
     cols = tl.arange(0, BLOCK_N)
